Add files via upload
XGBoost is a new Machine Learning algorithm designed with speed andperformance in mind. XGBoost stands for eXtreme Gradient Boosting andis based on decision trees. In this project, we will import theXGBClassifier from the xgboost library; this is an implementation of thescikit-learn API for XGBoost classification.
The model achieves an accuracy of 94.87 % , which is great considering the number of lines of code.